requirement.txt
mlflow==2.12.1
google-generativeai>=0.7.0
pandas>=1.5.0
dagshub>=0.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
protobuf>=3.20.0